Update README.md
To make the e-Store project work, you need to follow these steps for table migration:

Access the database and go to the "roles" table. Insert "customer" in the "name" column for a regular customer role and "admin" to log in as an administrator. Set "guard_name" to "web".
After registering, you can log in to the database if you want to make any user an administrator.

If you are logged in as an administrator, go to the dashboard and fill in the products and slides. For slides, the photo resolution should be 1920 x 450.